chore(log): address SonarCloud issues

+++ src/libcommon/utility/utility.h
@@ -516,7 +516,7 @@ struct COMMON_EXPORT CommonUtility {
          */
         static ExitInfo logDirectoryPath(SyncPath &directoryPath) noexcept;
 
-        static ExitInfo stdErrorToExitInfo(int error) noexcept;
+        static ExitInfo stdErrorToExitInfo(int64_t error) noexcept;
         static ExitInfo stdErrorToExitInfo(const std::error_code &ec) noexcept { return stdErrorToExitInfo(ec.value()); }
 
     private:
